feat (CI): check for cached DB instance

+++ ci/setup.sh
@@ -2,12 +2,12 @@
 
 set -e
 # check to see if exist folder is empty
-#if [ ! -d "$HOME/exist-$EXIST_DB_VERSION" ]; then
+if [ ! -d "$HOME/exist-$EXIST_DB_VERSION" ]; then
   wget https://github.com/eXist-db/exist/archive/${EXIST_DB_VERSION}.tar.gz -O /tmp/${EXIST_DB_VERSION}.tar.gz
   cd $HOME
   tar -xvf /tmp/${EXIST_DB_VERSION}.tar.gz
   cd exist-${EXIST_DB_VERSION}
   ./build.sh
-#else
-#  echo "Using cached database instance."
-#fi
+else
+  echo "Using cached database instance."
+fi
